+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added option to provide both the generator matrix and the check matrix to [`BlockCode`](https://komm.dev/ref/BlockCode), which must then satisfy $G H^\mathsf{T} = 0$.
 
+### Fix
+
+- Changed `BerlekampDecoder` to return the received word when the error locator has too few roots.
+
 ## v0.34.0 (2026-09-15)
 
 ### Added